Reorganize - Move primitive functions into actions.lua - Merge move() into actions.lua - Move base64.lua in /lib - Add belt_router.lua in /lib - Start work on event subscriber system
